% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_with_tz}
\alias{df_with_tz}
\title{Sets the timezone of all time-based columns in a dataframe}
\usage{
df_with_tz(df, tz = "UTC")
}
\arguments{
\item{df}{dataframe}

\item{tz}{timezone, an Olson timezone or "UTC" (default)}
}
\value{
dataframe
}
\description{
Sets the timezone of all time-based columns in a dataframe
}
\examples{
df_with_tz(wx_ames, tz = "UTC")

}

