% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph_pod.R, R/dygraph_pod_sidebar.R
\name{dygraph_server}
\alias{dygraph_server}
\alias{dygraph_sidebar_server}
\title{Server function for module that makes a dygraph.}
\usage{
dygraph_server(input, output, session, data)

dygraph_sidebar_server(input, output, session, data)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{data}{data.frame or a reactive that returns a data.frame}
}
\value{
\describe{
  \item{\code{ dygraph_server}}{a list containing:
    \itemize{
      \item \code{rct_result} a \code{shiny::\link[shiny]{reactive}},
        returning the resulting data.frame
      \item \code{rct_input_state} a \code{shiny::\link[shiny]{reactive}},
        returning a list of logicals describing the state of the inputs
      \item \code{rct_status_content} a \code{shiny::\link[shiny]{reactive}},
        returning a list with members \code{class} and \code{message} used to
        build the status output.
    }
  }
  \item{\code{ dygraph_sidebar_server}}{a reactive that returns a data.frame}
}
}
\description{
Used to define the server within the \code{ dygraph } shiny module.
}
\examples{
library("shiny")
shinyServer(function(input, output, session) {

  list_rct <- callModule(
    module = dygraph_sidebar_server,
    id = "foo",
    data = shinypod::wx_ames
  )

})

}
\seealso{
Other dygraph module functions: \code{\link{dygraph_sidebar_side}},
  \code{\link{dygraph_ui_input}},
  \code{\link{dygraph_ui_misc}},
  \code{\link{dygraph_ui_output}}
}

