% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph_pod_sidebar.R
\name{dygraph_sidebar_side}
\alias{dygraph_sidebar_main}
\alias{dygraph_sidebar_side}
\title{Sidebar layout for dygraph module}
\usage{
dygraph_sidebar_side(id)

dygraph_sidebar_main(id)
}
\arguments{
\item{id}{character, used to identify a namespace}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
These functions return the ui elements for a side panel and a main panel.
}
\details{
Generally, the side elements are the inputs; the main elements are the outputs.
}
\examples{
library("shiny")
shinyUI(
  fluidPage(
    shinyjs::useShinyjs(),
    sidebarLayout(
      sidebarPanel(dygraph_sidebar_side("foo")),
      mainPanel(dygraph_sidebar_main("foo"))
    )
  )
)

}
\seealso{
Other dygraph module functions: \code{\link{dygraph_server}},
  \code{\link{dygraph_ui_input}},
  \code{\link{dygraph_ui_misc}},
  \code{\link{dygraph_ui_output}}
}

