% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph.R
\name{dygraph_ui_input}
\alias{dygraph_ui_input}
\title{UI input elements for dygraph module.}
\usage{
dygraph_ui_input(id)
}
\arguments{
\item{id, }{character used to specify namesapce, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Used to define the UI input elements within the \code{dygraph} shiny module.
}
\details{
This function returns a \code{shiny::\link[shiny]{tagList}} with members:

\describe{
 \item{time}{\code{shiny::\link[shiny]{selectizeInput}}, used to specify time variable}
 \item{y1}{\code{shiny::\link[shiny]{selectizeInput}}, used to specify y1-axis variable}
 \item{y2}{\code{shiny::\link[shiny]{selectizeInput}}, used to specify y2-axis variable}
}

The purpose is to specify the UI elements - another set of functions can be used to specify layout.
}
\seealso{
Other dygraph module functions: \code{\link{dygraph_server}},
  \code{\link{dygraph_ui_misc}},
  \code{\link{dygraph_ui_output}}
}

