% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph.R
\name{dygraph_ui_misc}
\alias{dygraph_ui_misc}
\title{UI miscellaneous elements for dygraph module.}
\usage{
dygraph_ui_misc(id)
}
\arguments{
\item{id, }{character used to specify namesapce, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Used to define the UI input elements within the \code{dygraph} shiny module.
}
\details{
This function returns a \code{shiny::\link[shiny]{tagList}} with members:

\describe{
 \item{help}{\code{shiny::\link[shiny]{tags}$pre}, contains guidance for using dygraph}
}

The purpose is to specify the UI elements - another set of functions can be used to specify layout.
}
\seealso{
Other dygraph module functions: \code{\link{dygraph_server}},
  \code{\link{dygraph_ui_input}},
  \code{\link{dygraph_ui_output}}
}

