% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph_pod.R
\name{dygraph_ui_output}
\alias{dygraph_ui_output}
\title{UI output elements for module that makes a dygraph.}
\usage{
dygraph_ui_output(id)
}
\arguments{
\item{id, }{character used to specify namesapce, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Used to define the UI output elements within the \code{ dygraph } shiny module.
}
\details{
This function returns a \code{shiny::\link[shiny]{tagList}} with members:

\describe{
 \item{status}{\code{shiny::\link[shiny]{htmlOutput}},
   used to display status of the module}
}

The purpose is to specify the UI elements - another set of functions can be used to specify layout.
}
\seealso{
Other dygraph module functions: \code{\link{dygraph_server}},
  \code{\link{dygraph_sidebar_side}},
  \code{\link{dygraph_ui_input}},
  \code{\link{dygraph_ui_misc}}
}

