% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo_pod.R, R/foo_pod_sidebar.R
\name{foo_server}
\alias{foo_server}
\alias{foo_sidebar_server}
\title{Server function for does a foo.}
\usage{
foo_server(input, output, session, data, status_alert = TRUE)

foo_sidebar_server(input, output, session, data, status_show = TRUE,
  status_alert = TRUE)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{data}{data.frame, possibly reactive}

\item{status_alert}{logical, possibly reactive, indicates if to change alert-class of status output}

\item{status_show}{logical, possibly reactive, indicates if to show the status output}
}
\value{
\code{ foo_server}: a list containing reactives that return a data.frame
   and a list of logicals that describe the state of the module,
   \code{ foo_sidebar_server}: a reactive that returns a data.frame
}
\description{
Used to define the server within the \code{ foo } shiny module.
}
\examples{
shinyServer(function(input, output, session) {

  list_rct <- callModule(
    module = foo_server,
    id = "foo"
  )

})

}
\seealso{
foo_sidebar_main

Other foo module functions: \code{\link{foo_ui_input}},
  \code{\link{foo_ui_misc}}, \code{\link{foo_ui_output}}
}

