% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo_pod_sidebar.R
\name{foo_sidebar_side}
\alias{foo_sidebar_side}
\title{Sidebar layout for foo module}
\usage{
foo_sidebar_side(id)
}
\arguments{
\item{id}{character, used to identify a namespace}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
These functions return the ui elements for a side panel and a main panel.
}
\details{
Generally, the side elements are the inputs; the main elements are the outputs.
}
\examples{
shinyUI(
  fluidPage(
    shinyjs::useShinyjs(),
    sidebarLayout(
      sidebarPanel(foo_sidebar_side("foo")),
      mainPanel(foo_sidebar_main("foo"))
    )
  )
)

}

