% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo_pod.R
\name{foo_ui_input}
\alias{foo_ui_input}
\title{UI input elements for module that does a foo.}
\usage{
foo_ui_input(id)
}
\arguments{
\item{id, }{character used to specify namesapce, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Used to define the UI input elements within the \code{ foo } shiny module.
}
\details{
This function returns a \code{shiny::\link[shiny]{tagList}} with members:

\describe{
 \item{button_yell}{\code{shiny::\link[shiny]{fileInput}}, button to ivoke upper-case}
}

The purpose is to specify the UI elements - another set of functions can be used to specify layout.
}
\seealso{
Other foo module functions: \code{\link{foo_server}},
  \code{\link{foo_ui_misc}}, \code{\link{foo_ui_output}}
}

