% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactive.R
\name{isValidy}
\alias{isValidy}
\title{Evaluates the validity of a reactive expression}
\usage{
isValidy(x)
}
\arguments{
\item{x}{\code{shiny::\link[shiny]{reactive}} expression
to be evaluated by \code{shiny::\link[shiny]{req}}}
}
\value{
logical, indicating the validity of \code{x}
}
\description{
The difference between this function and \code{shiny::\link[shiny]{isTruthy}}
is that this function will always return a (\code{TRUE}/\code{FALSE}),
rather than get hung-up on an invalid reactive.
(test this ASAP to make sure I am not mis-remembering)
}
\details{
This function can be useful for use with
\code{shinyjs::\link[shinyjs]{toggleState}} and
\code{shinyjs::\link[shinyjs]{toggle}}, to control the state or visibility
of html elements according to the validity of some reactive expression.
}
\examples{
\dontrun{
library("shinyjs")

# code within a Shiny server function
toggle(id = "my_control", condition = isValidy(rct_data()))
}
}

