% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Module}
\alias{Module}
\alias{Pod}
\title{Construct a Module}
\usage{
Module(fn, ...)

Pod(fn, ...)
}
\arguments{
\item{fn}{module definition; <function>}

\item{...}{static data shared by all instances of this type <dot-dot-dot>}
}
\value{
\url{ShinyPod:type}
}
\description{
Construct a Module
}
\section{Fields}{

\describe{
\item{\code{.our}}{shared static data; <environment>}

\item{\code{.my}}{the module itself; <environment>}

\item{\code{bare_id}}{ID of the module instance, verbatim as supplied with; <character>}

\item{\code{parent_module}}{module in which this module is called; \url{ShinyPod:instance}}

\item{\code{parent_id}}{parent module's ID; <character>}

\item{\code{parent_ns}}{parent module's Shiny namespace; <function>}

\item{\code{id}}{ID of this instance , namespaced by parent \code{ns}; <character>}

\item{\code{ns}}{Shiny namespace of this module, itself namespaced by parent \code{ns}; <function>}

\item{\code{state}}{exposed state of the module; <reactivevalues>}

\item{\code{run}}{run the module a Shiny app; <function>}
}}

