% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{observe_class_swap}
\alias{observe_class_swap}
\title{swap out classes on an html element}
\usage{
observe_class_swap(id, expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{id}{A character vector to identify the html element to operate on.}

\item{expr}{An expression that returns a character vector to add to the html element.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
nothing
}
\description{
This function stores the value of the last class to be added (using this function),
then removes that class before addding the new class. For example, this may be useful
if you want to modify a panel to show an alert.
}
\details{
As this is an observer, there is no return value. It is called for the side-effect of
changing the class of the html element.

This is based on \code{shiny::renderText()}
}
