% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim_pod_sidebar.R
\name{read_delim_sb_side}
\alias{read_delim_sb_main}
\alias{read_delim_sb_side}
\title{Presentation UI-elements for read-delim module}
\usage{
read_delim_sb_side(id)

read_delim_sb_main(id)
}
\arguments{
\item{id}{character, id for the module}
}
\value{
\describe{
  \item{read_delim_sb_side}{
    \code{shiny::\link[shiny]{tagList}} of containing elements
    to put in a \code{shiny::\link[shiny]{sidebarPanel}}
  }
  \item{read_delim_sb_main}{
    \code{shiny::\link[shiny]{tagList}} of containing elements
    to put in a \code{shiny::\link[shiny]{mainPanel}}
  }
}
}
\description{
This is essentially a wrapper around
\code{readr::\link[readr]{read_delim}}.
These functions are used to create the UI elements for a
sidebar presentation of the read-delim module.
}
\details{
\describe{
  \item{read_delim_sb_side}{}
  \item{read_delim_sb_main}{}
}
}
\seealso{
\code{\link{read_delim_ui}}, \code{\link{read_delim_sb_server}}
}

