% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim.R, R/read_delim_sidebar.R
\name{read_delim_server}
\alias{read_delim_server}
\alias{read_delim_sidebar_server}
\title{Server function for delimted-file reader.}
\usage{
read_delim_server(input, output, session, delim = ",", decimal_mark = ".")

read_delim_sidebar_server(input, output, session, delim = ",",
  decimal_mark = ".")
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{delim}{character, default for parsing delimiter}

\item{decimal_mark}{character, default for decimal mark}
}
\value{
a list with members:
\code{rct_txt} \code{shiny::\link[shiny]{reactive}}, returns raw text
\code{rct_data} \code{shiny::\link[shiny]{reactive}}, returns tbl_df of the parsed text
}
\description{
Used to define the server within the \code{read_delim} shiny module.
}
\examples{
shinyServer(function(input, output, session) {

  list_rct <- callModule(
    module = read_delim_server,
    id = "foo"
  )

  observe(print(list_rct$rct_data()))
})

}
\seealso{
read_delim_sidebar_main

Other read_delim module functions: \code{\link{read_delim_ui_input}},
  \code{\link{read_delim_ui_misc}},
  \code{\link{read_delim_ui_output}}
}
\concept{read_delim module functions}
