% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim_pod_sidebar.R
\name{read_delim_sidebar_side}
\alias{read_delim_sidebar_main}
\alias{read_delim_sidebar_side}
\title{Sidebar layout for read_delim module}
\usage{
read_delim_sidebar_side(id)

read_delim_sidebar_main(id)
}
\arguments{
\item{id}{character, used to identify a namespace}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
These functions return the ui elements for a side panel and a main panel.
}
\details{
Generally, the side elements are the inputs; the main elements are the outputs.
}
\examples{
shinyUI(
  fluidPage(
    shinyjs::useShinyjs(),
    sidebarLayout(
      sidebarPanel(read_delim_sidebar_side("foo")),
      mainPanel(read_delim_sidebar_main("foo"))
    )
  )
)

}
\seealso{
Other read_delim module functions: \code{\link{read_delim_server}},
  \code{\link{read_delim_ui_input}},
  \code{\link{read_delim_ui_misc}},
  \code{\link{read_delim_ui_output}}
}

