% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim_pod.R
\name{read_delim_ui}
\alias{read_delim_ui}
\title{Creates the UI elements for read-delim Shiny module}
\usage{
read_delim_ui(id)
}
\arguments{
\item{id}{character, id for the module}
}
\value{
list with three elements:
\describe{
  \item{input}{\code{shiny::\link[shiny]{tagList}}
    containing Shiny inputs:
    \describe{
      \item{delim}{\code{shiny::\link[shiny]{selectInput}}
        delimiter, maps to \code{delim} arguement
      }
    }
  }
  \item{output}{\code{shiny::\link[shiny]{taglist}}
    containing Shiny outputs:
    \describe{
      \item{data_preview}{\code{shiny::\link[shiny]{htmlOutput}}
        used to display the first few lines of the parsed dataframe
      }
    }
  }
  \item{misc}{empty \code{shiny::\link[shiny]{tagList}}}
}
}
\description{
It is unlikely that you would call this function directly
from your Shiny app. Instead, this function would be called
by the functions you use to compose the UI of your Shiny app,
for example  \code{\link{read_delim_sb_side}} and
\code{\link{read_delim_sb_main}}.
}
\seealso{
\code{\link{read_delim_sb_side}},
  \code{\link{read_delim_sb_main}}, \code{\link{read_delim_server}}
}
\keyword{internal}

