% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim_pod.R
\name{read_delim_ui_input}
\alias{read_delim_ui_input}
\title{UI input elements for module that parses a delimited text file.}
\usage{
read_delim_ui_input(id)
}
\arguments{
\item{id, }{character used to specify namesapce, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Used to define the UI input elements within the \code{ read_delim } shiny module.
}
\details{
This function returns a \code{shiny::\link[shiny]{tagList}} with members:

\describe{
 \item{delim}{\code{shiny::\link[shiny]{selectizeInput}}, used to specify delimiter character}
 \item{decimal_mark}{\code{shiny::\link[shiny]{selectizeInput}}, used to specify decimal mark}
 \item{tz_parse}{\code{shiny::\link[shiny]{selectizeInput}}, used to specify timezone to parse}
 \item{tz_display}{\code{shiny::\link[shiny]{selectizeInput}}, used to specify timezone to display}
}

The purpose is to specify the UI elements - another set of functions can be used to specify layout.
}
\seealso{
Other read_delim module functions: \code{\link{read_delim_server}},
  \code{\link{read_delim_sidebar_side}},
  \code{\link{read_delim_ui_misc}},
  \code{\link{read_delim_ui_output}}
}

