% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim_pod.R
\name{read_delim_ui_misc}
\alias{read_delim_ui_misc}
\title{UI miscellaneous elements for module that parses a delimited text file.}
\usage{
read_delim_ui_misc(id)
}
\arguments{
\item{id, }{character used to specify namesapce, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Used to define the UI miscellaneous elements within the \code{ read_delim } shiny module.
}
\details{
This function returns a \code{shiny::\link[shiny]{tagList}} with members:

\describe{
}

The purpose is to specify the UI elements - another set of functions can be used to specify layout.
}
\seealso{
Other read_delim module functions: \code{\link{read_delim_server}},
  \code{\link{read_delim_sidebar_side}},
  \code{\link{read_delim_ui_input}},
  \code{\link{read_delim_ui_output}}
}

