% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim.R
\name{read_delim_ui_output}
\alias{read_delim_ui_output}
\title{UI output elements for delimited-file reader.}
\usage{
read_delim_ui_output(id)
}
\arguments{
\item{id, }{character used to specify namesapce, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Used to define the UI output elements within the \code{read_delim} shiny module.
}
\details{
This function returns a \code{shiny::\link[shiny]{tagList}} with members:

\describe{
 \item{text}{\code{shiny::\link[shiny]{htmlOutput}}, used to display first few lines of text from file}
 \item{data}{\code{shiny::\link[shiny]{htmlOutput}}, used to display first few lines of the parsed dataframe}
}

The purpose is to specify the UI elements - another set of functions can be used to specify layout.
}
\seealso{
Other read_delim.module.functions: \code{\link{read_delim_server}},
  \code{\link{read_delim_ui_input}}
}

