% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_delim_pod_sidebar.R
\name{sample_delim_sb_side}
\alias{sample_delim_sb_main}
\alias{sample_delim_sb_side}
\title{Presentation UI-elements for sample-delim module}
\usage{
sample_delim_sb_side(id)

sample_delim_sb_main(id)
}
\arguments{
\item{id}{character, id for the module}
}
\value{
\describe{
  \item{sample_delim_sb_side}{
    \code{shiny::\link[shiny]{tagList}} of containing elements
    to put in a \code{shiny::\link[shiny]{sidebarPanel}}
  }
  \item{sample_delim_sb_main}{
    \code{shiny::\link[shiny]{tagList}} of containing elements
    to put in a \code{shiny::\link[shiny]{mainPanel}}
  }
}
}
\description{
These functions are used to create the UI elements for a
sidebar presentation of the upload-text module.
}
\details{
\describe{
  \item{sample_delim_sb_side}{
    The only element in the sidebar panel is a
    \code{shiny::\link[shiny]{selectizeInput}}, used to
    select from some sample files that contain delimited text.
  }
  \item{sample_delim_sb_main}{
    The only element in the main panel is a
    \code{shiny::\link[shiny]{htmlOutput}}, used to
    display the first few lines of the uploaded file.
  }
}
}
\examples{
library("shiny")
library("shinyjs")
library("shinypod")

# also available by running sp_shiny_example("upload_text") // not yet

ui <- shinyUI(
  fluidPage(
    useShinyjs(),
    titlePanel("Select delimited-text file"),
    sidebarLayout(
      sidebarPanel(
        sample_delim_sb_side("sample_delim")
      ),
      mainPanel(
        sample_delim_sb_main("sample_delim")
      )
    )
  )
)

server <- shinyServer(function(input, output, session) {

  # rct_text is a reactive that can be used elsewhere in a server function
  rct_text <- callModule(module = sample_delim_sb_server, id = "sample_delim")

})

\dontrun{
  shinyApp(ui = ui, server = server)
}

}
\seealso{
\code{\link{sample_delim_ui}}, \code{\link{sample_delim_sb_server}}
}

