% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_delim_pod.R
\name{sample_delim_server}
\alias{sample_delim_server}
\title{Server function for sample-delim Shiny module}
\usage{
sample_delim_server(input, output, session)
}
\arguments{
\item{input}{\code{shiny::\link[shiny]{shinyServer}} input object}

\item{output}{\code{shiny::\link[shiny]{shinyServer}} output object}

\item{session}{\code{shiny::\link[shiny]{shinyServer}} session object}
}
\value{
\code{\link{sp_srv}} object with members:
\describe{
  \item{rct_result}{\code{shiny::\link[shiny]{reactive}},
    returns the character string read from the file
  }
  \item{rct_state}{\code{shiny::\link[shiny]{reactive}},
    returns list of logical values indicating the state
    of information available within the server function.
    This module has no "state".
  }
  \item{rct_notification}{\code{shiny::\link[shiny]{reactive}},
    returns list of arguments that can be passed to
    \code{shiny::\link[shiny]{showNotification}}:
    \describe{
      \item{ui}{character, content of the message}
      \item{type}{character, type of the message, i.e. "message"}
    }
  }
}
}
\description{
It is unlikely that you would call this function directly
from your Shiny app. Instead, this function would be called
by the server module that you call from your Shiny app,
for example  \code{\link{sample_delim_sb_server}}.
}
\details{
The component \code{rct_notification} remains \code{NULL} - no notifications.
}
\seealso{
\code{\link{sample_delim_sb_server}},
  \code{\link{sample_delim_ui}}
}
\keyword{internal}

