% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_delim_pod.R
\name{sample_delim_ui}
\alias{sample_delim_ui}
\title{Creates the UI elements for sample-delim Shiny module}
\usage{
sample_delim_ui(id)
}
\arguments{
\item{id}{character, id for the module}
}
\value{
list with three elements:
\describe{
  \item{input}{\code{shiny::\link[shiny]{tagList}}
    containing Shiny inputs:
    \describe{
      \item{select_delim}{\code{shiny::\link[shiny]{selectizeInput}}
        used to choose sample text
      }
    }
  }
  \item{output}{\code{shiny::\link[shiny]{taglist}}
    containing Shiny outputs:
    \describe{
      \item{text_preview}{\code{shiny::\link[shiny]{htmlOutput}}
        used to display the first few lines of the sampled
      }
    }
  }
  \item{misc}{empty \code{shiny::\link[shiny]{tagList}}}
}
}
\description{
It is unlikely that you would call this function directly
from your Shiny app. Instead, this function would be called
by the functions you use to compose the UI of your Shiny app,
for example  \code{\link{sample_delim_sb_side}} and
\code{\link{sample_delim_sb_main}}.
}
\seealso{
\code{\link{sample_delim_sb_side}},
  \code{\link{sample_delim_sb_main}}, \code{\link{sample_delim_server}}
}
\keyword{internal}

