% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compose.R
\name{sp_output.sp_ui}
\alias{sp_output.sp_ui}
\title{Access the Shiny output elements}
\usage{
\method{sp_output}{sp_ui}(sp_ui)
}
\arguments{
\item{sp_ui}{\code{\link{sp_ui}} object}
}
\value{
\code{shiny::\link[shiny]{tagList}} containing Shiny outputs
}
\description{
Access the Shiny output elements
}

