% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compose.R
\name{sp_rct_result.sp_srv}
\alias{sp_rct_result.sp_srv}
\title{Access the result reactive}
\usage{
\method{sp_rct_result}{sp_srv}(sp_srv)
}
\arguments{
\item{sp_srv}{\code{\link{sp_srv}} object}
}
\value{
\code{shiny::\link[shiny]{reactive}} that returns your result
}
\description{
Access the result reactive
}

