% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compose.R
\name{sp_rct_state.sp_srv}
\alias{sp_rct_state.sp_srv}
\title{Access the state reactive}
\usage{
\method{sp_rct_state}{sp_srv}(sp_srv)
}
\arguments{
\item{sp_srv}{\code{\link{sp_srv}} object}
}
\value{
\code{shiny::\link[shiny]{reactive}} that returns arguments
 that can be used with \code{shiny::\link[shiny]{showNotification}}
}
\description{
Access the state reactive
}

