% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_example.R
\name{sp_shiny_example}
\alias{sp_shiny_example}
\title{Runs (or lists) example shiny apps}
\usage{
sp_shiny_example(example_name = NULL)
}
\arguments{
\item{example_name}{character, name of example app to run}
}
\value{
character, list of available apps - or NULL (called for side effect)
}
\description{
Use this function to find out what example shiny-apps are available
with this package, or run any of them.
}
\examples{
sp_shiny_example()
\dontrun{
  sp_shiny_example("upload_text")
}
}

