% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compose.R
\name{sp_srv}
\alias{sp_srv}
\title{Constructs sp_srv object: template for server-return}
\usage{
sp_srv(rct_result, rct_state, rct_notification)
}
\value{
sp_srv object, S3 class
}
\description{
This function can be useful for composing your own
shinypods.
}
\details{
The idea of the foundation layer is to define the inputs, outputs, and how
they relate to each other. A sp_srv object has three members:

\describe{
  \item{rct_result}{A reactive that returns the
    result of the server function
  }
  \item{rct_state}{A reactive that returns a list of (presumably)
    logical values that can be used by the presentation server-function,
    for example, to show/hide various elements using
    \code{shinyjs::\link[shiny]{toggle}}.
  }
  \item{rct_notification}{A reactive that contains a list of arguements that
    the presentation server-function can use to construct a notification using
    \code{shiny::\link[shiny]{showNotification}}
  }
}

Presumably, the presentation server-function itself return \code{rct_result},
act upon the information in \code{rct_state} to show or hide UI elements,
and issue notifications based on \code{rct_notification}.
}
\seealso{
Accessor functions: \code{\link{sp_rct_result.sp_srv}},
  \code{\link{sp_rct_state.sp_srv}},
  \code{\link{sp_rct_notification.sp_srv}}
}

