% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compose.R
\name{sp_ui}
\alias{sp_ui}
\title{Constructs sp_ui object: template for UI elements}
\usage{
sp_ui()
}
\value{
sp_ui object, S3 class
}
\description{
This function can be useful for composing your own
shinypods.
}
\details{
As constructed, a sp_ui object is a list with three elements:

\describe{
  \item{input}{empty \code{shiny::\link[shiny]{tagList}}}
  \item{output}{empty \code{shiny::\link[shiny]{tagList}}}
  \item{misc}{empty \code{shiny::\link[shiny]{tagList}}}
}
}
\examples{
sp_ui()
}
\seealso{
Accessor functions: \code{\link{sp_input.sp_ui}}, \code{\link{sp_output.sp_ui}},
  \code{\link{sp_misc.sp_ui}}
}

