% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactive.R
\name{static}
\alias{static}
\title{Get the value of a reactive function, if reactive}
\usage{
static(x)
}
\arguments{
\item{x}{expression}
}
\value{
\code{x()} if reactive, \code{x} if not reactive
}
\description{
This is useful for functions where you want to be able to take either reactive
arguements or static arguments.
}
\keyword{internal}

