% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{status_content}
\alias{status_content}
\title{use input and result to generate message and class of status}
\usage{
status_content(status)
}
\arguments{
\item{status}{list with components \code{input} and \code{result}}
}
\value{
list with components \code{class} and \code{message}
}
\description{
The argument \code{status} shall be a list with two members: \code{input} and \code{result}.
Each of those lists shall have components \code{index}, \code{is_valid}, and \code{message}.
}
\details{
This return value is a list with members \code{class} and \code{message}. The \code{class} can be used by
\link{observe_class_swap} to change the appearance of an output. The \code{message} can be used as the
text displayed by the output.
}

