% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_html.R
\name{text_html}
\alias{text_html}
\title{Converts text into an HTML fragment}
\usage{
text_html(text, n_line = 6)
}
\arguments{
\item{text}{character, text with newline character}

\item{n_line}{numeric, number of lines of \code{text} to put into HTML}
}
\value{
chracter, HTML fragment
}
\description{
This is useful to compose an HTML fragment to put in an output built using
\code{shiny::\link[shiny]{renderUI}}. This is a useful way to show the first
few lines of a text file, verbatim.
}
\examples{
library("shiny")
library("readr")

ui <- shinyUI(

  fluidPage(
    sidebarLayout(
      sidebarPanel(),
      mainPanel(
        htmlOutput(outputId = "text_preview", container = pre_scroll)
      )
    )
  )
)

server <- shinyServer(function(input, output){

  output$text_preview <-
    renderUI(
      system.file("extdata", "wx_ames.csv", package = "shinypod") \%>\%
      read_file() \%>\%
      text_html()
    )

})

\dontrun{
shinyApp(ui = ui, server = server)
}

}
\seealso{
\code{\link{tibble_html}}, \code{\link{pre_scroll}}
}

