% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{tibble_html}
\alias{tibble_html}
\title{convert a tbl_df printout into an html fragment}
\usage{
tibble_html(data)
}
\arguments{
\item{data, }{data.frame}
}
\value{
html fragment
}
\description{
convert a tbl_df printout into an html fragment
}

