% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ui_server}
\alias{ui_server}
\alias{UI}
\alias{Server}
\title{Contruct a Shiny UI/Server Fragment}
\usage{
UI(fn)

Server(fn)
}
\arguments{
\item{fn}{Fragment definition; <function>}
}
\value{
\url{ShinyPod:ui_frag}/ShinyPod:server_frag>
}
\description{
Contruct a Shiny UI/Server Fragment
}
