% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_selected}
\alias{update_selected}
\title{determine the proper selection}
\usage{
update_selected(value, choices, index = NULL)
}
\arguments{
\item{value}{character vector, current value of an input}

\item{choices}{character vector, new choices for an input}

\item{index}{integer, if \code{value} is not in defualt \code{choices},
uses this index of \code{choices}.}
}
\value{
character vector of proposed selection
}
\description{
Used for \code{shiny::\link[shiny]{selectInput}} to allow you to
update its selection when its choices change.
}
\examples{
  update_selected("a", c("a", "b", "c"))
  update_selected("a", NULL)
  update_selected("d", c("a", "b", "c"))
  update_selected("d", c("a", "b", "c"), index = 1)

}
