% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_text_pod_sidebar.R
\name{upload_text_sb_server}
\alias{upload_text_sb_server}
\title{Presentation server-function for upload-text module}
\usage{
upload_text_sb_server(input, output, session)
}
\arguments{
\item{input}{\code{shiny::\link[shiny]{shinyServer}} input object}

\item{output}{\code{shiny::\link[shiny]{shinyServer}} output object}

\item{session}{\code{shiny::\link[shiny]{shinyServer}} session object}
}
\value{
\code{shiny::\link[shiny]{reactive}} that returns the
  character string read from the uploaded file
}
\description{
This is the server function for a
sidebar presentation of the upload-text module.
}
\details{
The return value is a reactive that returns, as a character string,
the contents of the uploaded file.

All the notifications provided by the foundation server-function,
\code{\link{upload_text_server}}, are passed along.
}
\examples{
library("shiny")
library("shinyjs")
library("shinypod")

# also available by running sp_shiny_example("upload_text")

ui <- shinyUI(
  fluidPage(
    useShinyjs(),
    titlePanel("Upload Text"),
    sidebarLayout(
      sidebarPanel(
        upload_text_sb_side("upload")
      ),
      mainPanel(
        upload_text_sb_main("upload")
      )
    )
  )
)

server <- shinyServer(function(input, output, session) {

  # rct_text is a reactive that can be used elsewhere in a server function
  rct_text <- callModule(module = upload_text_sb_server, id = "upload")

})

\dontrun{
  shinyApp(ui = ui, server = server)
}

}
\seealso{
\code{\link{upload_text_sb_side}}, \code{\link{upload_text_sb_main}},
  \code{\link{upload_text_server}}
}

