% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_text_pod_sidebar.R
\name{upload_text_sb_side}
\alias{upload_text_sb_main}
\alias{upload_text_sb_side}
\title{Presentation UI-elements for upload-text module}
\usage{
upload_text_sb_side(id)

upload_text_sb_main(id)
}
\arguments{
\item{id}{character, id for the module}
}
\value{
\describe{
  \item{upload_text_sb_side}{
    \code{shiny::\link[shiny]{tagList}} of containing elements
    to put in a \code{shiny::\link[shiny]{sidebarPanel}}
  }
  \item{upload_text_sb_main}{
    \code{shiny::\link[shiny]{tagList}} of containing elements
    to put in a \code{shiny::\link[shiny]{mainPanel}}
  }
}
}
\description{
These functions are used to create the UI elements for a
sidebar presentation of the upload-text module.
}
\details{
\describe{
  \item{upload_text_sb_side}{
    The only element in the sidebar panel is a
    \code{shiny::\link[shiny]{fileInput}}, used to
    upload a file from a local computer to the server.
  }
  \item{upload_text_sb_main}{
    The only element in the main panel is a
    \code{shiny::\link[shiny]{htmlOutput}}, used to
    display the first few lines of the uploaded file.
  }
}
}
\examples{
library("shiny")
library("shinyjs")
library("shinypod")

# also available by running sp_shiny_example("upload_text")

ui <- shinyUI(
  fluidPage(
    useShinyjs(),
    titlePanel("Upload Text"),
    sidebarLayout(
      sidebarPanel(
        upload_text_sb_side("upload")
      ),
      mainPanel(
        upload_text_sb_main("upload")
      )
    )
  )
)

server <- shinyServer(function(input, output, session) {

  # rct_text is a reactive that can be used elsewhere in a server function
  rct_text <- callModule(module = upload_text_sb_server, id = "upload")

})

\dontrun{
  shinyApp(ui = ui, server = server)
}

}
\seealso{
\code{\link{upload_text_ui}}, \code{\link{upload_text_sb_server}}
}

