% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_text.R, R/upload_text_sidebar.R
\name{upload_text_server}
\alias{upload_text_server}
\alias{upload_text_sidebar_server}
\title{Server function for delimted-file reader.}
\usage{
upload_text_server(input, output, session)

upload_text_sidebar_server(input, output, session)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}
}
\value{
\code{shiny::\link[shiny]{reactive}} that returns raw text
}
\description{
Used to define the server within the \code{upload_text} shiny module.
}
\examples{
shinyServer(function(input, output, session) {
})

}
\seealso{
upload_text_sidebar_main

Other upload_text module functions: \code{\link{upload_text_ui_input}},
  \code{\link{upload_text_ui_misc}},
  \code{\link{upload_text_ui_output}}
}

