% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_text_pod_sidebar.R
\name{upload_text_sidebar_side}
\alias{upload_text_sidebar_main}
\alias{upload_text_sidebar_side}
\title{Sidebar layout for upload_text module}
\usage{
upload_text_sidebar_side(id)

upload_text_sidebar_main(id)
}
\arguments{
\item{id}{character, used to identify a namespace}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
These functions return the ui elements for a side panel and a main panel.
}
\details{
Generally, the side elements are the inputs; the main elements are the outputs.
}
\examples{
shinyUI(
  fluidPage(
    shinyjs::useShinyjs(),
    sidebarLayout(
      sidebarPanel(upload_text_sidebar_side("foo")),
      mainPanel(upload_text_sidebar_main("foo"))
    )
  )
)

}
\seealso{
Other upload_text module functions: \code{\link{upload_text_server}},
  \code{\link{upload_text_ui_input}},
  \code{\link{upload_text_ui_misc}},
  \code{\link{upload_text_ui_output}}
}

