% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_text_pod.R
\name{upload_text_ui}
\alias{upload_text_ui}
\title{Creates the UI elements for upload-text Shiny module}
\usage{
upload_text_ui(id)
}
\arguments{
\item{id}{character, id for the module}
}
\value{
list with three elements:
\describe{
  \item{input}{\code{shiny::\link[shiny]{tagList}}
    containing Shiny inputs:
    \describe{
      \item{file}{\code{shiny::\link[shiny]{fileInput}}
        used to upload file
      }
    }
  }
  \item{output}{\code{shiny::\link[shiny]{taglist}}
    containing Shiny outputs:
    \describe{
      \item{text_preview}{\code{shiny::\link[shiny]{htmlOutput}}
        used to display the first few lines of the uploaded text file
      }
    }
  }
  \item{misc}{empty \code{shiny::\link[shiny]{tagList}}}
}
}
\description{
It is unlikely that you would call this function directly
from your Shiny app. Instead, this function would be called
by the functions you use to compose the UI of your Shiny app,
for example  \code{\link{upload_text_sb_side}} and
\code{\link{upload_text_sb_main}}.
}
\seealso{
\code{\link{upload_text_sb_side}},
  \code{\link{upload_text_sb_main}}, \code{\link{upload_text_server}}
}
\keyword{internal}

