% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{use_pod}
\alias{use_pod}
\title{create a shinypod from a template}
\usage{
use_pod(name, description, overwrite = FALSE)
}
\arguments{
\item{name}{character, name to prepend to the filenames}

\item{description}{character, short description to use in the function documentation}

\item{overwrite}{logical, indicates if an existing file can be overwritten}
}
\value{
list of TRUE values
}
\description{
This function is used to write out a template function for a shinypod,
both for the functional layer, in \code{foo_pod.R}, and a presentation layer, in
\code{foo_pod_sidebar.R}
}
\details{
This assumes that you are working in a directory in an R package; the files will be written to
the \code{R} directory.
}

