% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_delim.R
\name{write_delim_output}
\alias{write_delim_output}
\title{UI output elements for delimited-file writer.}
\usage{
write_delim_output(id)
}
\arguments{
\item{id, }{character used to specify namesapce, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Used to define the UI output elements within the \code{write_delim} shiny module.
}
\details{
This function returns a \code{shiny::\link[shiny]{tagList}} with members:

\describe{
 \item{data}{\code{shiny::\link[shiny]{htmlOutput}}, used to display first few lines of the dataframe}
 \item{text}{\code{shiny::\link[shiny]{htmlOutput}}, used to display first few lines of text from file}
 \item{text}{\code{shiny::\link[shiny]{htmlOutput}}, used to display first text status }
}

The purpose is to specify the UI elements - another set of functions can be used to specify layout.
}
\seealso{
Other write_delim module functions: \code{\link{write_delim_input}},
  \code{\link{write_delim_server}}
}

