% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_delim.R, R/write_delim_sidebar.R
\name{write_delim_server}
\alias{write_delim_server}
\alias{write_delim_sidebar_server}
\title{Server function for delimted-file writer.}
\usage{
write_delim_server(input, output, session, data, filename = "data.csv",
  delim = ",", status_alert = TRUE)

write_delim_sidebar_server(input, output, session, data, delim = ",",
  filename = "data.csv", status_show = TRUE, status_alert = TRUE)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{data}{data.frame}

\item{filename}{path}

\item{delim}{character, possibly reactive, delimiter mark to use as a default}

\item{status_alert}{logical, possibly reactive, indicates if to change alert-class of status output}

\item{status_show}{logical, possibly reactive, indicates if to show status}
}
\value{
a \code{shiny::\link[shiny]{reactive}} containing a tbl_df of the parsed text
}
\description{
Used to define the server within the \code{write_delim} shiny module.
}
\examples{
library("shiny")
shinyServer(function(input, output, session) {

  rct_data <- callModule(
    module = read_delim_server,
    id = "foo"
  )

  observe(print(rct_data()))
})

}
\seealso{
Other write_delim module functions: \code{\link{write_delim_ui_input}},
  \code{\link{write_delim_ui_output}}
}

