% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_delim_sidebar.R
\name{write_delim_sidebar_side}
\alias{write_delim_sidebar_main}
\alias{write_delim_sidebar_side}
\title{Sidebar layout for write_delim module}
\usage{
write_delim_sidebar_side(id)

write_delim_sidebar_main(id)
}
\arguments{
\item{id}{character, used to identify a namespace}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
These functions return the ui elements for a side panel and a main panel.
}
\details{
The side elements are the inputs; the main elements are the outputs.
}
\examples{
shinyUI(
  fluidPage(
    shinyjs::useShinyjs(),
    sidebarLayout(
      sidebarPanel(write_delim_sidebar("foo")),
      mainPanel(read_delim_main("foo"))
    )
  )
)

}

