% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_delim.R
\name{write_delim_ui_input}
\alias{write_delim_ui_input}
\title{UI input elements for delimited-file writer.}
\usage{
write_delim_ui_input(id)
}
\arguments{
\item{id, }{character used to specify namesapce, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Used to define the UI input elements within the \code{write_delim} shiny module.
}
\details{
This function returns a \code{shiny::\link[shiny]{tagList}} with members:

\describe{
 \item{delim}{\code{shiny::\link[shiny]{selectizeInput}}, used to specify delimiter character}
 \item{filename}{\code{shiny::\link[shiny]{textInput}}, used to specify file name}
 \item{download}{\code{shiny::\link[shiny]{downloadButton}}, download button}
}

The purpose is to specify the UI elements - another set of functions can be used to specify layout.
}
\seealso{
Other write_delim module functions: \code{\link{write_delim_server}},
  \code{\link{write_delim_ui_output}}
}

