% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_names_inherits}
\alias{df_names_inherits}
\title{Get the names of all the columns of the dataframe
that inherit from the supplied class name}
\usage{
df_names_inherits(data, what)
}
\arguments{
\item{data}{dataframe}

\item{what}{character, vector of class we wish to find}
}
\value{
character vector
}
\description{
Get the names of all the columns of the dataframe
that inherit from the supplied class name
}

