% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph.R
\name{dygraph_server}
\alias{dygraph_server}
\title{Server function for dygraph module.}
\usage{
dygraph_server(input, output, session, data)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{data}{data frame or \code{shiny::\link[shiny]{reactive}} that returns a data frame}
}
\value{
a \code{shiny::\link[shiny]{reactive}} that returns a dygraph
}
\description{
Used to define the server within the \code{dygraph} shiny module.
}
\examples{

}
\seealso{
Other dygraph module functions: \code{\link{dygraph_ui_input}},
  \code{\link{dygraph_ui_misc}},
  \code{\link{dygraph_ui_output}}
}

