% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph_sidebar.R
\name{dygraph_sidebar_side}
\alias{dygraph_sidebar_side}
\alias{dygraph_sidebar_main}
\title{Sidebar layout for dygraph module}
\usage{
dygraph_sidebar_side(id)

dygraph_sidebar_main(id, help = TRUE)
}
\arguments{
\item{id}{character, used to identify a namespace}

\item{help}{logical, indicates if help panels to be displayed}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
These functions return the ui elements for a side panel and a main panel.
}
\details{
The side elements are the inputs; the main elements are the outputs.
}
\examples{

}
