% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph.R
\name{dygraph_ui_output}
\alias{dygraph_ui_output}
\title{UI output elements for dygraph module.}
\usage{
dygraph_ui_output(id)
}
\arguments{
\item{id, }{character used to specify namesapce, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}}
}
\description{
Used to define the UI output elements within the \code{dygraph} shiny module.
}
\details{
Because there are no outputs,
this function returns an empty \code{shiny::\link[shiny]{tagList}}.

The purpose is to specify the UI elements - another set of functions can be used to specify layout.
}
\seealso{
Other dygraph module functions: \code{\link{dygraph_ui_input}},
  \code{\link{dygraph_ui_misc}}
}

