% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactive.R
\name{isValidy}
\alias{isValidy}
\title{Evaluates the validity of a reactive expression}
\usage{
isValidy(...)
}
\arguments{
\item{...}{expression to pass to \code{shiny::req()}}
}
\value{
logical, returns \code{TRUE} if shiny validation passes
}
\description{
Useful if you need to return \code{TRUE}/\code{FALSE} on the validity of a
shiny reactive expression
}

