% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pre_scroll}
\alias{pre_scroll}
\title{html for scrollable pre-formatted text}
\source{
\url{http://stackoverflow.com/questions/10374171/how-to-make-twitter-bootstraps-pre-blocks-scroll-horizontally}
}
\usage{
pre_scroll(...)
}
\arguments{
\item{...}{expression used to fill text}
}
\description{
This is used as the \code{container} argument in  \code{shiny::\link[shiny]{htmlOutput}}
}
