% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reactive_validate}
\alias{reactive_validate}
\title{combines handling of reactive and validating the contents}
\usage{
reactive_validate(expr, .f = identity, message, label = NULL, ...)
}
\arguments{
\item{expr}{expression, or reactive that returns an expressiondo}

\item{.f}{function that takes a single arg (object), returns TRUE if valid}

\item{message}{passed to need}

\item{label}{passed to need}

\item{...}{other args to pass along to .f}
}
\value{
reactive that returns the expression
}
\description{
combines handling of reactive and validating the contents
}

