% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim.R
\name{read_delim_server}
\alias{read_delim_server}
\title{Server function for delimted-file reader.}
\usage{
read_delim_server(input, output, session, delim = ",", decimal_mark = ".")
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{delim}{character, default for parsing delimiter}

\item{decimal_mark}{character, default for decimal mark}
}
\value{
a \code{shiny::\link[shiny]{reactive}} containing a tbl_df of the parsed text
}
\description{
Used to define the server within the \code{read_delim} shiny module.
}
\examples{
shinyServer(function(input, output, session) {

  rct_data <- callModule(
    module = read_delim_server,
    id = "foo"
  )

  observe(print(rct_data()))
})

}
\seealso{
Other read_delim module functions: \code{\link{dygraph_server}},
  \code{\link{read_delim_ui_input}},
  \code{\link{read_delim_ui_output}}
}

