% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim_pod.R, R/read_delim_pod_sidebar.R
\name{read_delim_server}
\alias{read_delim_server}
\alias{read_delim_sidebar_server}
\title{Server function for module that parses a delimited text file.}
\usage{
read_delim_server(input, output, session, text, delim = ",",
  decimal_mark = ".")

read_delim_sidebar_server(input, output, session, text, delim = ",",
  decimal_mark = ".")
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{delim}{character, defualt delimiter}

\item{decimal_mark}{character, default decimal-mark}

\item{data}{data.frame or a reactive that returns a data.frame}
}
\value{
\describe{
  \item{\code{ read_delim_server}}{a list containing:
    \itemize{
      \item \code{rct_result} a \code{shiny::\link[shiny]{reactive}},
        returning the resulting data.frame
      \item \code{rct_input_state} a \code{shiny::\link[shiny]{reactive}},
        returning a list of logicals describing the state of the inputs
      \item \code{rct_status_content} a \code{shiny::\link[shiny]{reactive}},
        returning a list with members \code{class} and \code{message} used to
        build the status output.
    }
  }
  \item{\code{ read_delim_sidebar_server}}{a reactive that returns a data.frame}
}
}
\description{
Used to define the server within the \code{ read_delim } shiny module.
}
\examples{
shinyServer(function(input, output, session) {

  list_rct <- callModule(
    module = read_delim_sidebar_server,
    id = "foo",
    data = iris
  )

})

}
\seealso{
Other read_delim module functions: \code{\link{read_delim_sidebar_side}},
  \code{\link{read_delim_ui_input}},
  \code{\link{read_delim_ui_misc}},
  \code{\link{read_delim_ui_output}}
}

