% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{text_html}
\alias{text_html}
\title{convert text into an html fragment}
\usage{
text_html(text, n = 6)
}
\arguments{
\item{text, }{text with newline character}

\item{n, }{number of lines to keep}
}
\value{
html fragment
}
\description{
convert text into an html fragment
}

