% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_html.R
\name{tibble_html}
\alias{tibble_html}
\title{Converts dataframe into an HTML fragment}
\usage{
tibble_html(data, n_row = 5)
}
\arguments{
\item{n_row}{numeric, number of rows of \code{df} to put into html}

\item{df}{data.frame, to display}
}
\value{
chracter, HTML fragment
}
\description{
This is useful to compose an HTML fragment to put in an output built using
\code{shiny::\link[shiny]{renderUI}}. This is a useful way to show the first
few rows of a dataframe, verbatim.
}
\examples{
library("shiny")
library("readr")

ui <- shinyUI(

  fluidPage(
    sidebarLayout(
      sidebarPanel(),
      mainPanel(
        htmlOutput(outputId = "text_preview", container = pre_scroll)
      )
    )
  )
)

server <- shinyServer(function(input, output){

  output$text_preview <-
    renderUI(
      system.file("extdata", "wx_ames.csv", package = "shinypod") \%>\%
      read_csv() \%>\%
      tibble_html()
    )

})

\dontrun{
shinyApp(ui = ui, server = server)
}
}
\seealso{
\code{\link{text_html}}, \code{\link{pre_scroll}}
}

