% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_text_pod.R, R/upload_text_pod_sidebar.R
\name{upload_text_server}
\alias{upload_text_server}
\alias{upload_text_sidebar_server}
\title{Server function for module that uploads a text file.}
\usage{
upload_text_server(input, output, session, data)

upload_text_sidebar_server(input, output, session, data)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{data}{data.frame or a reactive that returns a data.frame}
}
\value{
\describe{
  \item{\code{ upload_text_server}}{a list containing:
    \itemize{
      \item \code{rct_result} a \code{shiny::\link[shiny]{reactive}},
        returning the resulting text
      \item \code{rct_input_state} a \code{shiny::\link[shiny]{reactive}},
        returning a list of logicals describing the state of the inputs
      \item \code{rct_status_content} a \code{shiny::\link[shiny]{reactive}},
        returning a list with members \code{class} and \code{message} used to
        build the status output.
    }
  }
  \item{\code{ upload_text_sidebar_server}}{a reactive that returns a data.frame}
}
}
\description{
Used to define the server within the \code{ upload_text } shiny module.
}
\examples{
shinyServer(function(input, output, session) {

  list_rct <- callModule(
    module = upload_text_server,
    id = "foo",
    data = iris
  )

})

}
\seealso{
Other upload_text module functions: \code{\link{upload_text_sidebar_side}},
  \code{\link{upload_text_ui_input}},
  \code{\link{upload_text_ui_misc}},
  \code{\link{upload_text_ui_output}}
}

