% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_text_pod.R
\name{upload_text_ui_output}
\alias{upload_text_ui_output}
\title{UI output elements for module that uploads a text file.}
\usage{
upload_text_ui_output(id)
}
\arguments{
\item{id, }{character used to specify namesapce, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Used to define the UI output elements within the \code{ upload_text } shiny module.
}
\details{
This function returns a \code{shiny::\link[shiny]{tagList}} with members:

\describe{
 \item{status}{\code{shiny::\link[shiny]{htmlOutput}},
   used to display status of the module}
 \item{data_preview}{\code{shiny::\link[shiny]{htmlOutput}},
   used to display the first few rows of the dataframe}
}

The purpose is to specify the UI elements - another set of functions can be used to specify layout.
}
\seealso{
Other upload_text module functions: \code{\link{upload_text_server}},
  \code{\link{upload_text_sidebar_side}},
  \code{\link{upload_text_ui_input}},
  \code{\link{upload_text_ui_misc}}
}

