% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_delim.R
\name{write_delim_server}
\alias{write_delim_server}
\title{Server function for delimted-file writer.}
\usage{
write_delim_server(input, output, session, data, delim = ",",
  status_alert = TRUE)
}
\arguments{
\item{input}{standard \code{shiny} input}

\item{output}{standard \code{shiny} output}

\item{session}{standard \code{shiny} session}

\item{data}{data.frame}

\item{delim}{character, delimiter mark}
}
\value{
a \code{shiny::\link[shiny]{reactive}} containing a tbl_df of the parsed text
}
\description{
Used to define the server within the \code{write_delim} shiny module.
}
\examples{
shinyServer(function(input, output, session) {

  rct_data <- callModule(
    module = read_delim_server,
    id = "foo"
  )

  observe(print(rct_data()))
})

}
\seealso{
Other write_delim module functions: \code{\link{write_delim_ui_input}},
  \code{\link{write_delim_ui_output}}
}

